+++ internal/config/snapshot.go
@@ -6,5 +6,6 @@ import (
 
 type Snapshot struct {
 	ProposalsCheckInterval     time.Duration `env:"SNAPSHOT_PROPOSALS_CHECK_INTERVAL" envDefault:"1m"`
+	ProposalsUpdatesInterval   time.Duration `env:"SNAPSHOT_PROPOSALS_UPDATES_INTERVAL" envDefault:"10m"`
 	UnknownSpacesCheckInterval time.Duration `env:"SNAPSHOT_UNKNOWN_SPACES_CHECK_INTERVAL" envDefault:"1m"`
 }